@@ -14,10 +14,10 @@
     "url": "git+https://github.com/ygqygq2/workflow-wait-action.git"
   },
   "scripts": {
-    "prebuild": "rimraf dist",
+    "prebuild": "rimraf dist || true",
     "build": "ncc build src/index.ts --license LICENSES.txt && git add dist/",
     "lint": "pnpm lint:eslint",
-    "lint:eslint": "eslint src/**/*.{ts} --cache --fix",
+    "lint:eslint": "eslint src --cache --fix",
     "git:pre-commit": "pnpm build",
     "postinstall": "simple-git-hooks",
     "preinstall": "npx only-allow pnpm",
@@ -25,33 +25,27 @@
     "release": "bumpp"
   },
   "dependencies": {
-    "@actions/core": "^1.10.0",
-    "@actions/github": "^5.1.1"
+    "@actions/core": "^1.10.1",
+    "@actions/github": "^6.0.0"
   },
   "devDependencies": {
-    "@commitlint/config-conventional": "^17.4.2",
-    "@types/node": "^18.11.18",
-    "@typescript-eslint/eslint-plugin": "^5.50.0",
-    "@typescript-eslint/parser": "^5.50.0",
-    "@vercel/ncc": "^0.36.1",
-    "bumpp": "^8.2.1",
+    "@commitlint/config-conventional": "^19.1.0",
+    "@eslint/js": "^9.0.0",
+    "@types/node": "^20.12.6",
+    "@vercel/ncc": "^0.38.1",
+    "bumpp": "^9.4.0",
     "commitizen": "^4.3.0",
-    "cz-git": "^1.4.1",
-    "eslint": "^8.33.0",
-    "eslint-config-prettier": "^8.6.0",
-    "eslint-plugin-github": "^4.6.0",
-    "eslint-plugin-jest": "^27.2.1",
-    "eslint-plugin-prettier": "^4.2.1",
-    "eslint-plugin-zhlint": "^1.0.1",
-    "jest": "29.4.1",
+    "cz-git": "^1.9.1",
+    "eslint": "^9.0.0",
+    "jest": "29.7.0",
     "js-yaml": "^4.1.0",
-    "lint-staged": "^13.1.0",
-    "prettier": "^2.8.3",
-    "rimraf": "^4.1.2",
-    "simple-git-hooks": "^2.8.1",
-    "ts-jest": "^29.0.5",
-    "typescript": "^4.9.5",
-    "zhlint": "^0.7.1"
+    "lint-staged": "^15.2.2",
+    "prettier": "^3.2.5",
+    "rimraf": "^5.0.5",
+    "simple-git-hooks": "^2.11.1",
+    "ts-jest": "^29.1.2",
+    "typescript": "^5.4.4",
+    "typescript-eslint": "^7.6.0"
   },
   "simple-git-hooks": {
     "pre-commit": "pnpm lint-staged && pnpm git:pre-commit",
